+++ src/components/ui/Label/Label.stories.tsx
@@ -4,10 +4,18 @@
 import React from 'react';
 import Flex from '../Flex';
 import { Label } from './';
+import LabelDocs from './Label.mdx';
 
 export default {
   title: 'UI Components/Label',
+  parameters: {
+    docs: {
+      page: LabelDocs.parameters.docs.page().props.children.type
+    }
+  },
+  component: Label
 };
+
 
 export const BasicLabel = () => {
   return (